Remove GitHub repository "default branch" functionality & add "github-webhook" repository
The reason for this change is that the Terraform GitHub provider does not support adding new repositories and setting their default branch, for the reason that GitHub creates new repositories in an uninitialized state; when that happens, the "github_branch" resource fails to be created since no initial commit exists for the branch to reference. If the repository's "auto_init" flag is set to true (thus creating an initial commit), the "github_branch" resource fails because the branch already exists.

The solution for now is simply removing the default branch functionality altogether.

Additionally, this commit also adds a new "github-webhook" repository.