Disable KubeDNS Autoscaler "preventSinglePointFailure" flag
When this flag is on, it scales up KubeDNS to 2+ replicas as soon as
more than 1 node is detected. This is good to prevent single-point
failure, but will also prevent the cluster from going back to a single
node state (useful for low-cost cluster like ours).

Thus, adding a patch to the workflow to disable the flag in the KubeDNS
Autoscaler configuration.